move spec helpers to lib so that gem users can use them
