+++ lib/sharepoint.rb
@@ -1,2 +1,3 @@
 require 'sharepoint/errors'
 require 'sharepoint/client'
+require 'sharepoint/spec_helpers'
